+++ inject/inject-app/src/test/scala/com/twitter/inject/app/BindDSL.scala
@@ -57,8 +57,15 @@ private[twitter] trait BindDSL { self =>
    */
   def bindClass[T](clazz: Class[T]): ClassDSL[T] = new ClassDSL[T](clazz)
 
-  /** For Java Compatibility */
-  def bindClass[T](clazz: Class[T], instance: T): BindDSL.this.type = {
+  /**
+   * For Java Compatibility
+   *
+   * @note This method should be treated as `final` and NOT overridden, with
+   *       the exception of creating a Java-friendly forwarding method to ensure the
+   *       correct `self.type` type is returned for Java consumers. That method
+   *       should ONLY call `super`.
+   */
+  def bindClass[T](clazz: Class[T], instance: T): self.type = {
     addInjectionServiceModule(new TwitterModule {
       override def configure(): Unit = {
         bind(clazz).toInstance(instance)
@@ -67,8 +74,15 @@ private[twitter] trait BindDSL { self =>
     self
   }
 
-  /** For Java Compatibility */
-  def bindClass[T](clazz: Class[T], annotation: Annotation, instance: T): BindDSL.this.type = {
+  /**
+   * For Java Compatibility
+   *
+   * @note This method should be treated as `final` and NOT overridden, with
+   *       the exception of creating a Java-friendly forwarding method to ensure the
+   *       correct `self.type` type is returned for Java consumers. That method
+   *       should ONLY call `super`.
+   */
+  def bindClass[T](clazz: Class[T], annotation: Annotation, instance: T): self.type = {
     addInjectionServiceModule(new TwitterModule {
       override def configure(): Unit = {
         bind(clazz).annotatedWith(annotation).toInstance(instance)
@@ -77,8 +91,15 @@ private[twitter] trait BindDSL { self =>
     self
   }
 
-  /** For Java Compatibility */
-  def bindClass[T, Ann <: Annotation](clazz: Class[T], annotationClazz: Class[Ann], instance: T): BindDSL.this.type = {
+  /**
+   * For Java Compatibility
+   *
+   * @note This method should be treated as `final` and NOT overridden, with
+   *       the exception of creating a Java-friendly forwarding method to ensure the
+   *       correct `self.type` type is returned for Java consumers. That method
+   *       should ONLY call `super`.
+   */
+  def bindClass[T, Ann <: Annotation](clazz: Class[T], annotationClazz: Class[Ann], instance: T): self.type = {
     addInjectionServiceModule(new TwitterModule {
       override def configure(): Unit = {
         bind(clazz).annotatedWith(annotationClazz).toInstance(instance)
@@ -87,8 +108,15 @@ private[twitter] trait BindDSL { self =>
     self
   }
 
-  /** For Java Compatibility */
-  def bindClass[T, U <: T](clazz: Class[T], instanceClazz: Class[U]): BindDSL.this.type = {
+  /**
+   * For Java Compatibility
+   *
+   * @note This method should be treated as `final` and NOT overridden, with
+   *       the exception of creating a Java-friendly forwarding method to ensure the
+   *       correct `self.type` type is returned for Java consumers. That method
+   *       should ONLY call `super`.
+   */
+  def bindClass[T, U <: T](clazz: Class[T], instanceClazz: Class[U]): self.type = {
     addInjectionServiceModule(new TwitterModule {
       override def configure(): Unit = {
         bind(clazz).to(instanceClazz)
@@ -97,8 +125,15 @@ private[twitter] trait BindDSL { self =>
     self
   }
 
-  /** For Java Compatibility */
-  def bindClass[T, U <: T](clazz: Class[T], annotation: Annotation, instanceClazz: Class[U]): BindDSL.this.type = {
+  /**
+   * For Java Compatibility
+   *
+   * @note This method should be treated as `final` and NOT overridden, with
+   *       the exception of creating a Java-friendly forwarding method to ensure the
+   *       correct `self.type` type is returned for Java consumers. That method
+   *       should ONLY call `super`.
+   */
+  def bindClass[T, U <: T](clazz: Class[T], annotation: Annotation, instanceClazz: Class[U]): self.type = {
     addInjectionServiceModule(new TwitterModule {
       override def configure(): Unit = {
         bind(clazz).annotatedWith(annotation).to(instanceClazz)
@@ -107,8 +142,15 @@ private[twitter] trait BindDSL { self =>
     self
   }
 
-  /** For Java Compatibility */
-  def bindClass[T, Ann <: Annotation, U <: T](clazz: Class[T], annotationClazz: Class[Ann], instanceClazz: Class[U]): BindDSL.this.type = {
+  /**
+   * For Java Compatibility
+   *
+   * @note This method should be treated as `final` and NOT overridden, with
+   *       the exception of creating a Java-friendly forwarding method to ensure the
+   *       correct `self.type` type is returned for Java consumers. That method
+   *       should ONLY call `super`.
+   */
+  def bindClass[T, Ann <: Annotation, U <: T](clazz: Class[T], annotationClazz: Class[Ann], instanceClazz: Class[U]): self.type = {
     addInjectionServiceModule(new TwitterModule {
       override def configure(): Unit = {
         bind(clazz).annotatedWith(annotationClazz).to(instanceClazz)
@@ -120,7 +162,7 @@ private[twitter] trait BindDSL { self =>
   /* Private */
 
   private[app] class TypeDSL[T: TypeTag] {
-    def toInstance(instance: T): BindDSL.this.type = {
+    def toInstance(instance: T): self.type = {
       addInjectionServiceModule(new TwitterModule {
         override def configure(): Unit = {
           bind(asManifest[T]).toInstance(instance)
@@ -129,7 +171,7 @@ private[twitter] trait BindDSL { self =>
       self
     }
 
-    def to[U <: T: TypeTag]: BindDSL.this.type = {
+    def to[U <: T: TypeTag]: self.type = {
       addInjectionServiceModule(new TwitterModule {
         override def configure(): Unit = {
           bind(asManifest[T]).to(asManifest[U])
@@ -138,7 +180,7 @@ private[twitter] trait BindDSL { self =>
       self
     }
 
-    def to[U <: T: TypeTag](instanceClazz: Class[U]): BindDSL.this.type = {
+    def to[U <: T: TypeTag](instanceClazz: Class[U]): self.type = {
       addInjectionServiceModule(new TwitterModule {
         override def configure(): Unit = {
           bind(asManifest[T]).to(instanceClazz)
@@ -155,7 +197,7 @@ private[twitter] trait BindDSL { self =>
   }
 
   private[app] class TypeAnnotationDSL[T: TypeTag, Ann <: Annotation: TypeTag] extends TypeDSL[T] {
-    override def toInstance(instance: T): BindDSL.this.type = {
+    override def toInstance(instance: T): self.type = {
       addInjectionServiceModule(new TwitterModule {
         override def configure(): Unit = {
           bind(asManifest[T], asManifest[Ann]).toInstance(instance)
@@ -164,7 +206,7 @@ private[twitter] trait BindDSL { self =>
       self
     }
 
-    override def to[U <: T: TypeTag]: BindDSL.this.type = {
+    override def to[U <: T: TypeTag]: self.type = {
       addInjectionServiceModule(new TwitterModule {
         override def configure(): Unit = {
           bind(asManifest[T]).annotatedWith(asManifest[Ann]).to(asManifest[U])
@@ -173,7 +215,7 @@ private[twitter] trait BindDSL { self =>
       self
     }
 
-    override def to[U <: T: TypeTag](instanceClazz: Class[U]): BindDSL.this.type = {
+    override def to[U <: T: TypeTag](instanceClazz: Class[U]): self.type = {
       addInjectionServiceModule(new TwitterModule {
         override def configure(): Unit = {
           bind(asManifest[T]).annotatedWith(asManifest[Ann]).to(instanceClazz)
@@ -184,7 +226,7 @@ private[twitter] trait BindDSL { self =>
   }
 
   private[app] class TypeAnnotationClassDSL[T: TypeTag, Ann <: Annotation](annotationClazz: Class[Ann]) extends TypeDSL[T] {
-    override def toInstance(instance: T): BindDSL.this.type = {
+    override def toInstance(instance: T): self.type = {
       addInjectionServiceModule(new TwitterModule {
         override def configure(): Unit = {
           bind(asManifest[T]).annotatedWith(annotationClazz).toInstance(instance)
@@ -193,7 +235,7 @@ private[twitter] trait BindDSL { self =>
       self
     }
 
-    override def to[U <: T: TypeTag]: BindDSL.this.type = {
+    override def to[U <: T: TypeTag]: self.type = {
       addInjectionServiceModule(new TwitterModule {
         override def configure(): Unit = {
           bind(asManifest[T]).annotatedWith(annotationClazz).to(asManifest[U])
@@ -202,7 +244,7 @@ private[twitter] trait BindDSL { self =>
       self
     }
 
-    override def to[U <: T: TypeTag](instanceClazz: Class[U]): BindDSL.this.type = {
+    override def to[U <: T: TypeTag](instanceClazz: Class[U]): self.type = {
       addInjectionServiceModule(new TwitterModule {
         override def configure(): Unit = {
           bind(asManifest[T]).annotatedWith(annotationClazz).to(instanceClazz)
@@ -213,7 +255,7 @@ private[twitter] trait BindDSL { self =>
   }
 
   private[app] class TypeWithNamedAnnotationDSL[T: TypeTag](annotation: Annotation) extends TypeDSL[T] {
-    override def toInstance(instance: T): BindDSL.this.type = {
+    override def toInstance(instance: T): self.type = {
       addInjectionServiceModule(new TwitterModule {
         override def configure(): Unit = {
           bind(asManifest[T]).annotatedWith(annotation).toInstance(instance)
@@ -222,7 +264,7 @@ private[twitter] trait BindDSL { self =>
       self
     }
 
-    override def to[U <: T: TypeTag]: BindDSL.this.type = {
+    override def to[U <: T: TypeTag]: self.type = {
       addInjectionServiceModule(new TwitterModule {
         override def configure(): Unit = {
           bind(asManifest[T]).annotatedWith(annotation).to(asManifest[U])
@@ -231,7 +273,7 @@ private[twitter] trait BindDSL { self =>
       self
     }
 
-    override def to[U <: T: TypeTag](instanceClazz: Class[U]): BindDSL.this.type = {
+    override def to[U <: T: TypeTag](instanceClazz: Class[U]): self.type = {
       addInjectionServiceModule(new TwitterModule {
         override def configure(): Unit = {
           bind(asManifest[T]).annotatedWith(annotation).to(instanceClazz)
@@ -242,7 +284,7 @@ private[twitter] trait BindDSL { self =>
   }
 
   private[app] class ClassDSL[T](clazz: Class[T]) {
-    def toInstance(instance: T): BindDSL.this.type = {
+    def toInstance(instance: T): self.type = {
       addInjectionServiceModule(new TwitterModule {
         override def configure(): Unit = {
           bind(clazz).toInstance(instance)
@@ -251,7 +293,7 @@ private[twitter] trait BindDSL { self =>
       self
     }
 
-    def to[U <: T : TypeTag : ClassTag]: BindDSL.this.type = {
+    def to[U <: T : TypeTag : ClassTag]: self.type = {
       addInjectionServiceModule(new TwitterModule {
         override def configure(): Unit = {
           bind(clazz).to(typeLiteral[U](asManifest[U]))
@@ -260,7 +302,7 @@ private[twitter] trait BindDSL { self =>
       self
     }
 
-    def to[U <: T](instanceClazz: Class[U]): BindDSL.this.type = {
+    def to[U <: T](instanceClazz: Class[U]): self.type = {
       addInjectionServiceModule(new TwitterModule {
         override def configure(): Unit = {
           bind(clazz).to(instanceClazz)
@@ -275,7 +317,7 @@ private[twitter] trait BindDSL { self =>
   }
 
   private[app] class ClassAnnotationDSL[T, Ann <: Annotation](clazz: Class[T], annotationClazz: Class[Ann]) extends ClassDSL(clazz) {
-    override def toInstance(instance: T): BindDSL.this.type = {
+    override def toInstance(instance: T): self.type = {
       addInjectionServiceModule(new TwitterModule {
         override def configure(): Unit = {
           bind(clazz).annotatedWith(annotationClazz).toInstance(instance)
@@ -284,7 +326,7 @@ private[twitter] trait BindDSL { self =>
       self
     }
 
-    override def to[U <: T : TypeTag : ClassTag]: BindDSL.this.type = {
+    override def to[U <: T : TypeTag : ClassTag]: self.type = {
       addInjectionServiceModule(new TwitterModule {
         override def configure(): Unit = {
           bind(clazz).annotatedWith(annotationClazz).to(typeLiteral[U](asManifest[U]))
@@ -293,7 +335,7 @@ private[twitter] trait BindDSL { self =>
       self
     }
 
-    override def to[U <: T](instanceClazz: Class[U]): BindDSL.this.type = {
+    override def to[U <: T](instanceClazz: Class[U]): self.type = {
       addInjectionServiceModule(new TwitterModule {
         override def configure(): Unit = {
           bind(clazz).annotatedWith(annotationClazz).to(instanceClazz)
@@ -304,7 +346,7 @@ private[twitter] trait BindDSL { self =>
   }
 
   private[app] class ClassWithNamedAnnotationDSL[T](clazz: Class[T], annotation: Annotation) extends ClassDSL(clazz) {
-    override def toInstance(instance: T): BindDSL.this.type = {
+    override def toInstance(instance: T): self.type = {
       addInjectionServiceModule(new TwitterModule {
         override def configure(): Unit = {
           bind(clazz).annotatedWith(annotation).toInstance(instance)
@@ -313,7 +355,7 @@ private[twitter] trait BindDSL { self =>
       self
     }
 
-    override def to[U <: T : TypeTag : ClassTag]: BindDSL.this.type = {
+    override def to[U <: T : TypeTag : ClassTag]: self.type = {
       addInjectionServiceModule(new  TwitterModule {
         override def configure(): Unit = {
           bind(clazz).annotatedWith(annotation).to(typeLiteral[U](asManifest[U]))
@@ -323,7 +365,7 @@ private[twitter] trait BindDSL { self =>
     }
 
 
-    override def to[U <: T](instanceClazz: Class[U]): BindDSL.this.type = {
+    override def to[U <: T](instanceClazz: Class[U]): self.type = {
       addInjectionServiceModule(new TwitterModule {
         override def configure(): Unit = {
           bind(clazz).annotatedWith(annotation).to(instanceClazz)