+++ inject/inject-app/src/test/scala/com/twitter/inject/app/EmbeddedApp.scala
@@ -3,6 +3,7 @@ package com.twitter.inject.app
 import collection.JavaConverters._
 import com.google.inject.Module
 import com.twitter.inject.{Injector, Logging}
+import java.lang.annotation.Annotation
 import scala.annotation.varargs
 
 /**
@@ -48,6 +49,30 @@ class EmbeddedApp(app: com.twitter.inject.app.App) extends BindDSL with Logging 
 
   def main(flags: java.util.Map[String, Any]): Unit =
     main(flags = flags.asScala.toMap)
+
+  // java-forwarder methods
+  override final def bindClass[T](clazz: Class[T], instance: T): this.type =
+    super.bindClass[T](clazz, instance)
+
+  // java-forwarder methods
+  override final def bindClass[T](clazz: Class[T], annotation: Annotation, instance: T): this.type =
+    super.bindClass[T](clazz, annotation, instance)
+
+  // java-forwarder methods
+  override final def bindClass[T, Ann <: Annotation](clazz: Class[T], annotationClazz: Class[Ann], instance: T): this.type =
+    super.bindClass[T, Ann](clazz, annotationClazz, instance)
+
+  // java-forwarder methods
+  override final def bindClass[T, U <: T](clazz: Class[T], instanceClazz: Class[U]): this.type =
+    super.bindClass[T, U](clazz, instanceClazz)
+
+  // java-forwarder methods
+  override final def bindClass[T, U <: T](clazz: Class[T], annotation: Annotation, instanceClazz: Class[U]): this.type =
+    super.bindClass[T, U](clazz, annotation, instanceClazz)
+
+  // java-forwarder methods
+  override final def bindClass[T, Ann <: Annotation, U <: T](clazz: Class[T], annotationClazz: Class[Ann], instanceClazz: Class[U]): this.type =
+    super.bindClass[T, Ann, U](clazz, annotationClazz, instanceClazz)
 
   /* Protected */