+++ inject/inject-app/src/test/scala/com/twitter/inject/app/TestInjector.scala
@@ -4,7 +4,9 @@ import com.google.inject.{Module, Stage}
 import com.twitter.app.{Flag, FlagParseException, FlagUsageError, Flags}
 import com.twitter.inject.Injector
 import com.twitter.inject.app.internal.Modules
+import java.lang.annotation.Annotation
 import java.util.concurrent.atomic.AtomicBoolean
+import scala.annotation.varargs
 import scala.collection.JavaConverters._
 
 /**
@@ -32,10 +34,9 @@ object TestInjector {
    * @param modules - a variable list of [[com.google.inject.Module]]
    * @return a new [[TestInjector]]
    *
-   * @note Java users should prefer `apply(java.util.Collection<Module>)`
    * @see https://twitter.github.io/finatra/user-guide/testing/index.html#integration-tests
    */
-  def apply(modules: Module*): TestInjector = {
+  @varargs def apply(modules: Module*): TestInjector = {
     apply(modules = modules)
   }
 
@@ -125,6 +126,30 @@ class TestInjector(
     start()
     underlying
   }
+
+  // java-forwarder methods
+  override final def bindClass[T](clazz: Class[T], instance: T): this.type =
+    super.bindClass[T](clazz, instance)
+
+  // java-forwarder methods
+  override final def bindClass[T](clazz: Class[T], annotation: Annotation, instance: T): this.type =
+    super.bindClass[T](clazz, annotation, instance)
+
+  // java-forwarder methods
+  override final def bindClass[T, Ann <: Annotation](clazz: Class[T], annotationClazz: Class[Ann], instance: T): this.type =
+    super.bindClass[T, Ann](clazz, annotationClazz, instance)
+
+  // java-forwarder methods
+  override final def bindClass[T, U <: T](clazz: Class[T], instanceClazz: Class[U]): this.type =
+    super.bindClass[T, U](clazz, instanceClazz)
+
+  // java-forwarder methods
+  override final def bindClass[T, U <: T](clazz: Class[T], annotation: Annotation, instanceClazz: Class[U]): this.type =
+    super.bindClass[T, U](clazz, annotation, instanceClazz)
+
+  // java-forwarder methods
+  override final def bindClass[T, Ann <: Annotation, U <: T](clazz: Class[T], annotationClazz: Class[Ann], instanceClazz: Class[U]): this.type =
+    super.bindClass[T, Ann, U](clazz, annotationClazz, instanceClazz)
 
   /* Protected */