+++ inject/inject-server/src/test/scala/com/twitter/inject/server/EmbeddedTwitterServer.scala
@@ -10,6 +10,7 @@ import com.twitter.inject.modules.InMemoryStatsReceiverModule
 import com.twitter.inject.server.PortUtils.getPort
 import com.twitter.util.lint.{GlobalRules, Rule}
 import com.twitter.util.{Await, Closable, Duration, ExecutorServiceFuturePool, Future}
+import java.lang.annotation.Annotation
 import java.util.concurrent.ConcurrentLinkedQueue
 import java.util.concurrent.atomic.AtomicBoolean
 import org.scalatest.Matchers
@@ -280,6 +281,36 @@ class EmbeddedTwitterServer(
   lazy val adminHostAndPort: String = PortUtils.loopbackAddressForPort(httpAdminPort())
 
   /* Public */
+
+  // java-forwarder methods - DO NOT override except for creating a java-forwarder method
+  // that only calls `super()`
+  override def bindClass[T](clazz: Class[T], instance: T): this.type =
+    super.bindClass[T](clazz, instance)
+
+  // java-forwarder methods - DO NOT override except for creating a java-forwarder method
+  // that only calls `super()`
+  override def bindClass[T](clazz: Class[T], annotation: Annotation, instance: T): this.type =
+    super.bindClass[T](clazz, annotation, instance)
+
+  // java-forwarder methods - DO NOT override except for creating a java-forwarder method
+  // that only calls `super()`
+  override def bindClass[T, Ann <: Annotation](clazz: Class[T], annotationClazz: Class[Ann], instance: T): this.type =
+    super.bindClass[T, Ann](clazz, annotationClazz, instance)
+
+  // java-forwarder methods - DO NOT override except for creating a java-forwarder method
+  // that only calls `super()`
+  override def bindClass[T, U <: T](clazz: Class[T], instanceClazz: Class[U]): this.type =
+    super.bindClass[T, U](clazz, instanceClazz)
+
+  // java-forwarder methods - DO NOT override except for creating a java-forwarder method
+  // that only calls `super()`
+  override def bindClass[T, U <: T](clazz: Class[T], annotation: Annotation, instanceClazz: Class[U]): this.type =
+    super.bindClass[T, U](clazz, annotation, instanceClazz)
+
+  // java-forwarder methods - DO NOT override except for creating a java-forwarder method
+  // that only calls `super()`
+  override def bindClass[T, Ann <: Annotation, U <: T](clazz: Class[T], annotationClazz: Class[Ann], instanceClazz: Class[U]): this.type =
+    super.bindClass[T, Ann, U](clazz, annotationClazz, instanceClazz)
 
   /**
    * Returns the result of running the `nonExitingMain` of the underlying TwitterServer in